+++ test/org/ojalgo/matrix/store/ColumnsCase.java
@@ -25,24 +25,28 @@ import org.junit.jupiter.api.BeforeEach;
 import org.ojalgo.random.Uniform;
 import org.ojalgo.scalar.ComplexNumber;
 
-public class SelectedRowsCase extends NonPhysicalTest {
+public class ColumnsCase extends NonPhysicalTest {
 
+    @Override
     @BeforeEach
     public void setUp() {
 
-        final int tmpRowDim = Uniform.randomInteger(1, 9);
-        final int tmpColDim = Uniform.randomInteger(1, 9);
+        int tmpRowDim = Uniform.randomInteger(1, 9);
+        int tmpColDim = Uniform.randomInteger(1, 9);
 
-        final MatrixStore<ComplexNumber> tmpBase = NonPhysicalTest.makeRandomMatrix(tmpRowDim, tmpColDim);
+        MatrixStore<ComplexNumber> tmpBase = NonPhysicalTest.makeRandomMatrix(tmpRowDim, tmpColDim);
 
-        final int[] tmpRows = new int[Uniform.randomInteger(1, tmpRowDim)];
-        for (int i = 0; i < tmpRows.length; i++) {
-            tmpRows[i] = Uniform.randomInteger(tmpRowDim);
+        int[] tmpCols = new int[Uniform.randomInteger(1, tmpColDim)];
+        for (int i = 0; i < tmpCols.length; i++) {
+            tmpCols[i] = Uniform.randomInteger(tmpColDim);
         }
 
-        rationalStore = new RowsStore<>(GenericDenseStore.RATIONAL.copy(tmpBase), tmpRows);
-        complexStore = new RowsStore<>(GenericDenseStore.COMPLEX.copy(tmpBase), tmpRows);
-        primitiveStore = new RowsStore<>(PrimitiveDenseStore.FACTORY.copy(tmpBase), tmpRows);
+        rationalStore = new ColumnsStore<>(GenericDenseStore.RATIONAL.copy(tmpBase), tmpCols);
+        complexStore = new ColumnsStore<>(GenericDenseStore.COMPLEX.copy(tmpBase), tmpCols);
+        primitiveStore = new ColumnsStore<>(PrimitiveDenseStore.FACTORY.copy(tmpBase), tmpCols);
+
+        numberOfRows = tmpRowDim;
+        numberOfColumns = tmpCols.length;
     }
 
 }